+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 2.5.5
+* Fix bug duplicate permission Android
+
+## 2.5.4
+* Fix bug custom permission Android
+* Remove calling notification when task remove
+* Fixed calling notification show
+* Fixed some bugs.
+
+## 2.5.3
+* Update Android 14+ compatibility(force CallStyle/`FOREGROUND_SERVICE_PHONE_CALL` for https://developer.android.com/about/versions/14/behavior-changes-all#non-dismissable-notifications)
+* Change structure Notification/Sound on Android
+* Update get avatar from assets(Android), thanks @Ricky-yu https://github.com/hiennguyen92/flutter_callkit_incoming/pull/674
+* Add `isAccepted` properties for iOS and only show facetime button if support video, thank @td2thinh https://github.com/hiennguyen92/flutter_callkit_incoming/pull/673
+* Fixed some bugs.
+
 ## 2.5.3-alpha
 * Update Android 14+ compatibility(force CallStyle/`FOREGROUND_SERVICE_PHONE_CALL` for https://developer.android.com/about/versions/14/behavior-changes-all#non-dismissable-notifications)
 * Fixed some bugs.